hidpi: fix issue with text size vs GDK_SCALE
Disabling *.dpi in Xresources, and setting GDK_SCALE in env fixes the
inconsistent sizing issues I've been having, where icons, ui elements
and some text would be to small.

Additionally, fix the font size in alacritty, was set low, I assumed it
was because of the dpi setting, but I found that there.