zsh/login: add MOZ_ENABLE_WAYLAND
Don't think this does anything anymore, but I'll keep it here for now,
so that isn't the issue at least.

Context: having issues with firefox-developer-edition. I run a 4k
display, with scaling set to 2. This works very well for all native
wayland applications except for firefox-developer-edition for some
ungodly reason, fonts are blurry and pixelated, just like it usually is
in an Xwayland session but it's not running in Xwayland. I don't know
what's wrong or how to fix it though, and setting this flag is about all
I can think of doing for now.