hypr: mark default ws as persistent
In an attempt to fix the wonky monitor assignment issue I've been having
where workspaces will move to a different monitor on wakeup. I belive it
has to do with my KVM switch, not necessarily hyprland issue as such,
hoping this will mitigate. Didn't have this issue on sway though.