hypr: rework monitor config
Have multiple layouts, and need to be able to switch between them
easily. ~/.config/hypr/monitor/current.conf is a symlink, and points to
whatever ./scripts/switch-display.sh sets it to. I track the symlink for
now, but might remove later depending on new host.

There are still several issues but good enough to commit.