hypr: fix issue with passing mouse btn to hunt
Mouse 275 is bound to kando normally, but I use the key for games as
well, and I need to be able to disable it easily. Ideally I'd like to
match on class rules for the global shortcut, but that doesn't work.
Best I can do is pass the shortcut to hunt, but kando will still respond
so remember to temporarily disable it.