quickshell: add basic bar
https://quickshell.outfoxxed.me/docs/guide/introduction/

Mostly following that example, but adding some basic appearance options,
like a palette and font variables. Palette is gonna need to be replaced
with upstream, but works as an example for now.

Untracked ags in the process, not going to work on that further, might
later, but I'll re-add it then.