hypr: fix issue with crash on monitor change
So I had this persistent issue where some apps would crash on either
monitor layout switch, or from DPMS wakeup. Not all apps, which clued me
to that it's xwayland apps that is crashing. The culprit is
force_zero_scaling, and for whatever reason having this on, then
changing monitor settings cause apps to crash, presumably something to
do with wayland vs xwayland scaling. I just have to live with pixelated
  xwayland apps for now.