hypr,zsh: re-enable GDK_SCALE + xwayland scaling
The issue is that it kills apps on monitor layout/dpms change. I
narrowed it down to xwayland apps, and setting the scaling option seems
to mitigate this somewhat. But when not forcing scaling, there are a
multitude of issues running various games, I'm left with no option than
to run gamescope to get native resolution, but this in turn causes
frame time issues sometimes. This is the best of two bad options, and
hopefully this becomes better with time.