qs/alerts: refactor services + add audio in alert
Pipewire is used for more than audio handling, so it made sense to stick
that in its own service. AudioOut works, but there is no way I can find
to tell if the stream is actively playing, but there's a lot of data
missing. Not really something I need to alert about though, so I left it
not visible for now.