qs/hypr: fix activeToplevel sync
There were instances where the active top level would not update
correctly, always on reboot, and hyprland startup, but it would
sometimes happen in a way I was unable to reproduce yet.

This fixes the boot issue at least, and I suspect if fixes the other
issue as well, because the urgent status was always correct, it was just
that active top level didn't match.